Authentication and Authorization

+++ src/main/java/com/stayHub/stayHub/service/HotelServiceImpl.java
@@ -5,12 +5,15 @@ import com.stayHub.stayHub.dto.HotelInfoDto;
 import com.stayHub.stayHub.dto.RoomDto;
 import com.stayHub.stayHub.entity.Hotel;
 import com.stayHub.stayHub.entity.Room;
+import com.stayHub.stayHub.entity.User;
 import com.stayHub.stayHub.exception.ResoureceNotFoundException;
+import com.stayHub.stayHub.exception.UnAuthorisedExceptions;
 import com.stayHub.stayHub.repositry.HotelRepository;
 import com.stayHub.stayHub.repositry.RoomRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -31,6 +34,10 @@ public class HotelServiceImpl implements HotelService{
         log.info("Creating a new hotel with name: {}" , hotelDto.getName());
         Hotel hotel = modelMapper.map(hotelDto, Hotel.class);
         hotel.setActive(false);
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        hotel.setOwner(user);
+
         hotel = hotelRepository.save(hotel);
         log.info("Created a new hotel with ID: {}" , hotelDto.getId());
         return modelMapper.map(hotel , HotelDto.class);
@@ -42,6 +49,12 @@ public class HotelServiceImpl implements HotelService{
         Hotel hotel = hotelRepository
                 .findById(id)
                 .orElseThrow( ()-> new ResoureceNotFoundException("Hotel Not found with ID : " + id));
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (! user.equals(hotel.getOwner())) {
+            throw new UnAuthorisedExceptions("This user does not ow this hotel with id: " + id);
+        }
+
         return modelMapper.map(hotel, HotelDto.class);
     }
 
@@ -51,6 +64,12 @@ public class HotelServiceImpl implements HotelService{
         Hotel hotel = hotelRepository
                 .findById(id)
                 .orElseThrow( ()-> new ResoureceNotFoundException("Hotel is not found with id :" + id));
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (! user.equals(hotel.getOwner())) {
+            throw new UnAuthorisedExceptions("This user does not ow this hotel with id: " + id);
+        }
+
         modelMapper.map(hotelDto, hotel);
         hotel.setId(id);
         hotel = hotelRepository.save(hotel);
@@ -64,6 +83,11 @@ public class HotelServiceImpl implements HotelService{
         Hotel hotel = hotelRepository
                 .findById(id)
                         .orElseThrow( ()-> new ResoureceNotFoundException("Hotel not found with ID: " + id));
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (! user.equals(hotel.getOwner())) {
+            throw new UnAuthorisedExceptions("This user does not ow this hotel with id: " + id);
+        }
 
         for(Room room : hotel.getRooms()){
             inventoryService.deleteAllInventories(room);
@@ -79,7 +103,14 @@ public class HotelServiceImpl implements HotelService{
         Hotel hotel = hotelRepository
                 .findById(hotelId)
                 .orElseThrow( ()-> new ResoureceNotFoundException("Hotel is not found with id :" + hotelId));
-       hotel.setActive(true);
+
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (! user.equals(hotel.getOwner())) {
+            throw new UnAuthorisedExceptions("This user does not ow this hotel with id: " + hotelId);
+        }
+
+        hotel.setActive(true);
 
         // assuming only do it once
 
@@ -89,6 +120,7 @@ public class HotelServiceImpl implements HotelService{
 
     }
 
+    // Publie method
     @Override
     public HotelInfoDto getHotelInfoById(Long hotelId) {
